% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cutoff.R
\name{get_cutoff}
\alias{get_cutoff}
\title{Get min/max cutoff value for a performance metric of the notNMD model}
\usage{
get_cutoff(value, metric, model = "base")
}
\arguments{
\item{value}{value for cutoff}

\item{metric}{what performance metric the value relates to}

\item{model}{use base or lncRNA model?}
}
\value{
min or max (or both) cutoff value where the metric is above the value
}
\description{
Get min/max cutoff value for a performance metric of the notNMD model
}
\examples{
get_cutoff(0.9, "Pos Pred Value")
get_cutoff(0.9, "Neg Pred Value")
get_cutoff(0.8, "Accuracy")
}
\author{
Beth Signal
}
