% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictNMD.R
\name{predictNMD}
\alias{predictNMD}
\title{Predict if a transcript is subject to nonsense mediated decay}
\usage{
predictNMD(orfs, output = "class", model = "base")
}
\arguments{
\item{orfs}{orf information data.frame produced by GeneStructureTools::getOrfs()}

\item{output}{class or prob - output a class prediction or probability of nonsense mediated decay?}

\item{model}{use base or lncRNA model?}
}
\value{
vector with predicted class or probaility for nonsense mediated decay
}
\description{
Predict if a transcript is subject to nonsense mediated decay
}
\author{
Beth Signal
}
