% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictNMD.R
\name{predictNMD}
\alias{predictNMD}
\title{Predict if a transcript is subject to nonsense mediated decay}
\usage{
predictNMD(orfs, output = "class")
}
\arguments{
\item{orfs}{orf information data.frame produced by GeneStructureTools::get_orfs()}
}
\value{
vector with predicted class
}
\description{
Predict if a transcript is subject to nonsense mediated decay
}
\author{
Beth Signal
}
